added a start of cli;   can do: build from csv, build from gui, show full, solve
